requirements.txt
streamlit
opencv-python-headless==4.8.1.78
mediapipe
numpy
Pillow